refactor(lib): move key generators out into a util/keys module
