+++ src/lib/vec/index.test.ts
@@ -1,4 +1,4 @@
-import { indicesByKey, swizzledKeys, vec2, vec3, vec4 } from './index';
+import { vec2, vec3, vec4 } from './index';
 
 describe('vec/index', () => {
   test('vec2, vec3, vec4', () => {
@@ -38,15 +38,5 @@ describe('vec/index', () => {
     stpq.s = 1;
     stpq[1] = 2;
     expect([].slice.call(stpq.st)).toEqual([1, 2]);
-  });
-
-  test('config', () => {
-    expect({
-      indicesByKey,
-      swizzledKeys,
-    }).toMatchSnapshot();
-
-    expect(indicesByKey.size).toBe(4 * 3);
-    expect(swizzledKeys.size).toBe((4 + 4 ** 2 + 4 ** 3 + 4 ** 4) * 3);
   });
 });
