+++ src/lib/vec/index.ts
@@ -1,61 +1,13 @@
-// const { isArray } = Array;
+import { swizzledKeys, indicesByKey } from './util/keys';
+
 const { get, set } = Reflect;
-const { concat, slice } = [];
+const { slice } = [];
 
-const baseKeys = [
-  ['x', 'y', 'z', 'w'],
-  ['s', 't', 'p', 'q'],
-  ['r', 'g', 'b', 'a'],
-];
-
-function toSwizzled(
-  swizzled: string[],
-  _: string,
-  i: number,
-  keys: string[],
-): string[] {
-  const toSwizzledDepth = (swizzledDepth: string[], key: string): string[] =>
-    swizzledDepth.concat(keys.map(k => `${key}${k}`));
-
-  return i > 0 ? swizzled.reduce(toSwizzledDepth, keys) : swizzled;
+function toArray(acc: number[], arg: Component): number[] {
+  return acc.concat(typeof arg === 'number' ? arg : slice.call(arg));
 }
 
-export const swizzledKeys = new Set(
-  baseKeys.reduce(
-    (acc: string[], keys: string[]) => [
-      ...acc,
-      ...keys.reduceRight(toSwizzled, keys),
-    ],
-    [],
-  ),
-);
-
-function toIndicesByKey(
-  acc: Array<[string, number]>,
-  key: string,
-  i: number,
-): Array<[string, number]> {
-  return [...acc, [key, i]];
-}
-
-export const indicesByKey = new Map(
-  baseKeys.reduce(
-    (acc: Array<[string, number]>, keys: string[]) => [
-      ...acc,
-      ...keys.reduce(toIndicesByKey, []),
-    ],
-    [],
-  ),
-);
-
-function flatten(args: Array<Component | Components>): ArrayLike<number> {
-  return concat.apply(
-    [],
-    args.map(arg => (typeof arg === 'number' ? arg : slice.call(arg))),
-  );
-}
-
-function getSwizzled(target: IVector, prop: string): Component | Components {
+function getSwizzled(target: Vector, prop: string): Component {
   if (prop.length === 1) {
     return target[(indicesByKey.has(prop) && indicesByKey.get(prop)) as number];
   }
@@ -71,13 +23,9 @@ function getSwizzled(target: IVector, prop: string): Component | Components {
   );
 }
 
-function setSwizzled(
-  target: IVector,
-  prop: string,
-  value: Component | Components,
-) {
+function setSwizzled(target: Vector, prop: string, value: Component) {
   const keys = prop.split('');
-  const values = flatten([value]);
+  const values = [value].reduce(toArray, []);
 
   if (keys.length > values.length) {
     throw new Error('not enough data provided for construction');
@@ -95,22 +43,22 @@ function setSwizzled(
   return true;
 }
 
-const handler: ProxyHandler<IVector> = {
-  get(target: IVector, prop: PropertyKey) {
+const handler: ProxyHandler<Vector> = {
+  get(target: Vector, prop: PropertyKey) {
     return typeof prop === 'string' && swizzledKeys.has(prop)
       ? getSwizzled(target, prop)
       : get(target, prop);
   },
 
-  set(target: IVector, prop: PropertyKey, value: Component | Components) {
+  set(target: Vector, prop: PropertyKey, value: Component) {
     return typeof prop === 'string' && swizzledKeys.has(prop)
       ? setSwizzled(target, prop, value)
       : set(target, prop, value);
   },
 };
 
-function vec(size: number, args: Array<Component | Components>): IVector {
-  const components = flatten(args);
+function vec(size: number, args: Components): Vector {
+  const components = args.reduce(toArray, []);
 
   if (components.length < size) {
     throw new Error('not enough data provided for construction');
@@ -120,7 +68,7 @@ function vec(size: number, args: Array<Component | Components>): IVector {
     throw new Error('too many arguments');
   }
 
-  return new Float32Array(components) as IVector;
+  return new Float32Array(components) as Vector;
 }
 
 const vecs = new Array(4)
@@ -130,8 +78,7 @@ const vecs = new Array(4)
       i === 0
         ? acc
         : acc.concat(
-            (...args: Array<Component | Components>) =>
-              new Proxy(vec(i + 1, args), handler),
+            (...args: Components) => new Proxy(vec(i + 1, args), handler),
           ),
     [],
   );
